Update install_snort.sh
fixing conflicts, download now work but still some issues

+++ install_snort.sh
@@ -8,14 +8,15 @@
 #####################################################################
 
 
-
+#depedencies
 required_packages=('cmake' 'libdaq-dev' 'libdnet-dev' 'flex' 'g++' 'hwloc' 'libluajit-5.1-dev' 'libssl-dev' 'libpcap-dev' 'libpcre3-dev' 'pkg-config' 'zlib1g-dev')
 
 #default path for snort3 repository and libdaq library
-snort3_repository_path="~/"
-libdaq_repository_path="~/"
+snort3_repository_path="/home/cryptx/snort3"
+libdaq_repository_path="/home/cryptx/libdaq"
 libdaq_lib_path="/usr/local/lib/daq_s3"
 
+#colors
 GREEN="\033[0;32m"
 RED="\033[0;31m"
 NC="\033[0m"
@@ -34,7 +35,7 @@ install_depedencies() {
 }
 
 install_libdaq() {
-    cd "${libdaq_repository_path}"
+    cd /home/cryptx
     if ! git clone -q https://github.com/snort3/libdaq.git; then 
         echo -e "${RED}[✘] Error : couldn't get libdaq${NC}"
         return 0
@@ -42,35 +43,34 @@ install_libdaq() {
 
     cd libdaq
     ./bootstrap
-    echo "${libdaq_lib_path}"
-    ./configure --prefix="${libdaq_lib_path}"
-    make install
+    ./configure --prefix=/usr/local/lib/daq_s3
+    make install 
     sudo ldconfig
 
     echo -e "${GREEN}[✔] Libdaq installed !${NC}" 
     return 1
 }
 
-#args :
+# args :
 # $1 (optional) = path_install_snort
-# If you don't indicates manually the path, the default one will be used. Which is ~/
+# If you don't indicates manually the path, the default one will be used.
 install_snort3() {
-    cd "${snort3_repository_path}"
+    cd /home/cryptx
     if ! git clone -q https://github.com/snort3/snort3.git; then
         echo -e "${RED}[✘] Error : couldn't get snort3${NC}"
         exit 1
     fi
 
-    export snort3_installation="${snort3_repository_path}"
-    mkdir -p $snort3_installation
+    mkdir -p "${snort3_repository_path}"
     cd snort3
-    ./configure_cmake.sh --prefix="${snort3_installation}"
+    ./configure_cmake.sh --prefix="${snort3_repository_path}"
 
     cd build
     make -j $(nproc)
     make install
 
-    $"{snort3_installation}"bin/snort -V
+    "${snort3_repository_path}"/bin/snort -V
+    "${snort3_repository_path}"/bin/snort --daq-list
 
     echo -e "${GREEN}[✔] Snort3 installed !${NC}" 
     return 1
@@ -103,9 +103,20 @@ install() {
 }
 
 remove() {
-    sudo rm -rf "${snort3_repository_path}"
-    sudo rm -rf "${libdaq_lib_path}"
-    echo -e "${GREEN}[✔] Snort3 correcly removed !${NC}" 
+    if [ -d "${snort3_repository_path}" ]; then
+        sudo rm -rf "${snort3_repository_path}"
+        echo -e "${GREEN}[✔] Snort3 repository correcly removed !${NC}"
+    fi
+
+    if [ -d "${libdaq_repository_path}" ]; then
+        sudo rm -rf "${libdaq_repository_path}"
+        echo -e "${GREEN}[✔] Libdaq repository correcly removed !${NC}"
+    fi
+
+    if [ -d "${libdaq_lib_path}" ]; then
+        sudo rm -rf "${libdaq_lib_path}"
+        echo -e "${GREEN}[✔] Libdaq library correcly removed !${NC}"
+    fi
 }
 
 main() {
